requirements.txt
flask
huggingface_hub
python-dotenv
Pillow
